workaround for potential bug, if insert_order gets too big to fit long int

+++ lab2-scheduler/scheduler.h
@@ -82,7 +82,7 @@ private:
     priority_queue<Process*, vector<Process*>, PriorityComparator> *expired_queue;
     priority_queue<Process*, vector<Process*>, PriorityComparator> *temp;
     int quantum;
-    int insert_order;
+    long int insert_order;
 public:
     PRIO(int);
     Process* getEvent();
